requirements.txt
pygame
numpy
PyYAML